--- checkitbaby/fortigate_agent_ping.py
@@ -1,23 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Feb 12, 2020
-@author: cgustave
-"""
-import logging as log
-
-# create logger
-log.basicConfig(
-    format='%(asctime)s,%(msecs)3.3d %(levelname)-8s[%(module)-10.10s.\
-    %(funcName)-20.20s:%(lineno)5d] %(message)s',
-    datefmt='%Y%m%d:%H:%M:%S',
-    filename='debug.log',
-    level=log.DEBUG)
-    
-
-class Mixin:
-    """
-    Specific Fortigate agent ping functions loaded in FortiGate_agent
-    """
-
-    def process_ping(self, line=""):
-        log.info("Enter with line={}".format(line))
